requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
DAWG-Python==0.7.2
docopt==0.6.2
Flask==3.0.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
nltk==3.8.1
pyaes==1.6.1
PyJWT==2.8.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
SQLAlchemy==2.0.29
TgCrypto==1.2.5
tqdm==4.66.2
typing_extensions==4.10.0
Werkzeug==3.0.1
yarl==1.9.4
